+++ Hadrons/Modules/MDistil/LapEvec.hpp
@@ -29,10 +29,46 @@
 #ifndef Hadrons_MDistil_LapEvec_hpp_
 #define Hadrons_MDistil_LapEvec_hpp_
 
-#include <Hadrons/Modules/MDistil/Distil.hpp>
+#include <Hadrons/Global.hpp>
+#include <Hadrons/Module.hpp>
+#include <Hadrons/ModuleFactory.hpp>
+#include <Hadrons/EigenPack.hpp>
+#include <Hadrons/NamedTensor.hpp>
 
 BEGIN_HADRONS_NAMESPACE
 BEGIN_MODULE_NAMESPACE(MDistil)
+
+/*************************************************************************************
+ Rotate eigenvectors into our phase convention
+ First component of first eigenvector is real and positive
+ *************************************************************************************/
+
+inline void RotateEigen(std::vector<LatticeColourVector> & evec)
+{
+    ColourVector cv0;
+    auto grid = evec[0].Grid();
+    Coordinate siteFirst(grid->Nd(),0);
+    peekSite(cv0, evec[0], siteFirst);
+    const std::complex<Real> cplx0{cv0()()(0).real(), cv0()()(0).imag()};
+    if( cplx0.imag() == 0 )
+        LOG(Message) << "RotateEigen() : Site 0 : " << cplx0 << " => already meets phase convention" << std::endl;
+    else
+    {
+        const Real cplx0_mag{ std::abs(cplx0) };
+        const std::complex<Real> std_phase{std::conj(cplx0/cplx0_mag)};
+        LOG(Message) << "RotateEigen() : Site 0 : |" << cplx0 << "|=" << cplx0_mag
+                     << " => phase=" << (std::arg(std_phase) / M_PI) << " pi" << std::endl;
+        {
+            const Grid::Complex phase{std_phase.real(),std_phase.imag()};
+            for( int k = 0 ; k < evec.size() ; k++ )
+                evec[k] *= phase;
+            // Get rid of the rounding error in imaginary phase on the very first site
+            peekSite(cv0, evec[0], siteFirst);
+            cv0()()(0).imag(0); // this should be zero after the phase multiply - force it to be so
+            pokeSite(cv0, evec[0], siteFirst);
+        }
+    }
+}
 
 /******************************************************************************
